requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
notion-client==2.2.1
resend==2.4.0
python-dotenv==1.0.1
apscheduler==3.10.4
jinja2==3.1.4
